# server/docs/swagger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "tubu.io API Documentation - ",
    "version": "0.2.0",
    "title": "TUBU.IO API",
    "contact": {
      "email": "dev@tubu.io"
    }
  },
...
  "paths": {
...
    "/api/network": {
      "get": {
        "tags": ["Network"],
        "summary": "Get current user's active networks",
        "description": "",
        "operationId": "getNetworks",
        "parameters": [
          {
            "$ref": "#/components/parameters/pageFilter"
          },
          {
            "$ref": "#/components/parameters/pageSizeFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "consensus": {
                          "type": "string"
                        },
                        "bctype": {
                          "type": "string"
                        },
                        "version": {
                          "type": "string"
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "totalCount": {
                          "type": "integer"
                        },
                        "totalPage": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
    "/api/network/{networkID}": {
      "get": {
        "tags": ["Network"],
        "summary": "Get detail information of specific network",
        "description": "",
        "operationId": "getNetwork",
        "parameters": [
          {
            "$ref": "#/components/parameters/networkID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "allOf": [
                        { "$ref": "#/components/schemas/BasicNetwork" },
                        {
                          "type": "object",
                          "properties": {
                            "chain": {
                              "type": "object",
                              "properties": {
                                "blockCount": { "type": "number" },
                                "nodeInfo": { "type": "string" },
                                "chainId": { "type": "number" }
                              }
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      },
      "put": {
        "tags": ["Network"],
        "summary": "Update basic network information",
        "description": "",
        "operationId": "updateNetwork",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Network ID",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "Basic Network Information",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/BasicNetwork"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "405": {
            "description": "Invalid parameters"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
...
    "/api/transactions/": {
      "get": {
        "tags": ["Network", "Transaction"],
        "summary": "Get the network or contract transactions with the given networkID/shortID",
        "description": "",
        "operationId": "getTransactions",
        "parameters": [
          {
            "in": "query",
            "name": "networkID",
            "description": "Network ID",
            "required": false,
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "appID",
            "description": "Network ID",
            "required": false,
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "shortID",
            "description": "Short ID for contract",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/pageFilter"
          },
          {
            "$ref": "#/components/parameters/pageSizeFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "short_id": { "type": "string" },
                          "hash": { "type": "string" },
                          "from_address": { "type": "string" },
                          "status": { "type": "string" }
                        }
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
    "/api/transactions/{hash}": {
      "get": {
        "tags": ["Network", "Transaction"],
        "summary": "Get the transaction with the given transaction hash",
        "description": "",
        "operationId": "getTransactionsFromHash",
        "parameters": [
          {
            "in": "path",
            "name": "hash",
            "description": "Transaction Hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "short_id": {
                          "type": "string"
                        },
                        "from_address": {
                          "type": "string"
                        },
                        "hash": {
                          "type": "string"
                        },
                        "block_hash": {
                          "type": "string"
                        },
                        "input": {
                          "type": "object"
                        },
                        "status": {
                          "type": "string"
                        },
                        "block_number": {
                          "type": "string"
                        },
                        "extra_data": {
                          "type": "object"
                        },
                        "updated_at": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "error": "object",
            "message": "string"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "UserToken": []
          }
        ]
      }
    },
...
  },
  "security": [
    {
      "UserToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "UserToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
    "schemas": {
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "pageCount": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "rowCount": { "type": "integer" }
        },
        "xml": {
          "name": "MetaData"
        }
      },
...
      "BasicNetwork": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sample Network"
          },
          "description": {
            "type": "string"
          },
          "consensus": {
            "type": "string",
            "example": "raft",
            "description": "Consensus Algorithm"
          },
          "bctype": {
            "type": "string",
            "example": "quorum"
          },
          "version": {
            "type": "string",
            "example": "2.6.0"
          }
        },
        "xml": {
          "name": "Network"
        }
      },
...
    },
...
    "parameters": {
      "networkID": {
        "in": "path",
        "name": "networkID",
        "description": "Network ID",
        "schema": {
          "minimum": 1,
          "default": 1,
          "type": "integer"
        }
      },
...
      "pageFilter": {
        "in": "query",
        "name": "page",
        "description": "Active Page Number",
        "schema": {
          "minimum": 1,
          "default": 1,
          "type": "integer"
        }
      },
      "pageSizeFilter": {
        "in": "query",
        "name": "pageSize",
        "description": "Item count on a page",
        "schema": {
          "minimum": 1,
          "default": 20,
          "type": "integer"
        }
      },
...
    }
  }
}